Resolve #876: a numeric type is derived from the data, and nothing coerces it
`float(Int)` is `Float64`, so one integer argument inside a `promote_type` widened
the whole computation to `Float64` and a `Float32` caller lost its precision with
no error. The wrapper also decided for every number type the library has never
seen. It stood at more than a hundred sites over thirty-four files, and it spread
by copy: a new file took the line from its neighbour.

Every `float` wrapper under `src/` and `ext/` is gone. The type now comes from the
arguments through `eltype`, `typeof`, `real` and `promote_type`, and the arithmetic
widens it.

Where the operation is what widens, the type comes from the operation.
`cross_sectional_regression` answers a weighted least squares, so its working type
is `typeof(one(Ts) / one(Ts))`. That is wider than `float`, not narrower: an
integer panel still regresses in `Float64`, a `Float32` panel stays in `Float32`,
and an exact type stays exact.

Four coercions outside `float` go the same way.

  - `block_vec_pq` and the hierarchical clustering divided two integers with `/`
    and converted the quotient back. Both now divide with `÷`.
  - `ew_beta_output` built a `Matrix{Float64}` of a panel field. It now keeps the
    field's own type.
  - `factor_summary_ratio` typed its `NaN` on the numerator. It now types it on
    the quotient, which is the value it stands in for.

The rule is written in `.github/instructions/julia-source-code.instructions.md`
§ *Numeric types come from the data*, which is its Authority, and `STANDARDS.md`
carries the row that routes to it. `test/test_55_numeric_coercion_census.jl` is the
Gate: it reds on a `float` coercion under `src/` or `ext/`, and on an index that is
rounded in one type and converted in another.

Two testsets pinned the behaviour this commit abolishes, and both now pin the new
rule instead. `cross_sectional_cap_weights` of an integer panel raised to an
integer power stays integer, a fractional power lands in a float, and a `Float32`
panel stays `Float32`.

+++ src/21_FactorAttribution.jl
@@ -897,7 +897,7 @@ Means and variances scale by `ppy`, and volatilities by its square root. Shares 
 """
 function attribution_scale(ppy::Number)
     @argcheck(ppy > zero(ppy), DomainError(ppy, "ppy must be positive"))
-    return (; s1 = float(ppy), s2 = sqrt(float(ppy)))
+    return (; s1 = ppy, s2 = sqrt(ppy))
 end
 """
     attribution_idiosyncratic_matrix(esigma::VecNum)
@@ -949,7 +949,7 @@ A factor whose standalone volatility is zero has no correlation with anything, a
 """
 function attribution_safe_corr(cv::Number, s1::Number, s2::Number)
     d = s1 * s2
-    return d > zero(d) ? cv / d : convert(typeof(float(cv / oneunit(d))), NaN)
+    return d > zero(d) ? cv / d : convert(typeof(cv / oneunit(d)), NaN)
 end
 """
     attribution_family_index(fam::VecStr)
